MVP-project/pyproject.toml
[tool.poetry]
name = "mvp-project"
version = "0.1.0"
description = ""
authors = ["Andrey8705 <Andrey.shadrin.1998@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
pydantic = {extras = ["email"], version = "^2.10.4"}
bcrypt = "^4.2.1"
uvicorn = "^0.34.0"
python-jose = "^3.3.0"
uuid = "^1.30"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
